# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Invoice & Receipt System API
  version: 0.1.0
  description: |
    Internal API for a small Israeli business operating as `עוסק פטור`.

    Conventions that hold everywhere:

    - Money is an **integer number of agorot** (1 ILS = 100 agorot). Never a
      decimal, never a float.
    - Timestamps are RFC 3339 in UTC. Business dates are `YYYY-MM-DD`, already
      interpreted in `Asia/Jerusalem`.
    - Authentication is a session cookie. Unsafe methods must also send the
      `X-CSRF-Token` header, echoing the `csrf_token` cookie.
    - Errors always use one envelope. `code` is a stable English identifier;
      `message` is Hebrew and is shown to the user as-is.

    Sending a document never affects the document: delivery runs in a
    background job, after issuance has committed, so a failed send cannot roll
    back an issued document (plan.md 14).

    Critical commands accept an `Idempotency-Key` header. A replay returns the
    original response with `Idempotent-Replay: true`, so a retry after a timeout
    never issues a second document with a second official number.
servers:
  - url: /api/v1
...
security:
  - sessionCookie: []
paths:
...
  /payments/:
    get:
      tags: [payments]
      summary: List payments, newest first
      parameters:
        - name: customer_id
          in: query
          schema: { type: string, format: uuid }
        - name: activity_id
          in: query
          schema: { type: string, format: uuid }
        - name: method
          in: query
          schema: { $ref: '#/components/schemas/PaymentMethod' }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - name: unreceipted
          in: query
          schema: { type: boolean, default: false }
          description: Only payments that have no receipt yet
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of payments
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentPage' }
    post:
      tags: [payments]
      summary: Record a payment and issue its receipt
      description: |
        The primary flow of plan.md 12, in one transaction: validate, persist the
        payment, record what it settles, allocate the receipt number atomically,
        create the immutable receipt, render and store its PDF, audit, commit.
        The payment and its receipt can never disagree, because they commit
        together or not at all.

        Set `issue_receipt: false` to record the payment alone. That exists for
        the case where money arrives while the PDF renderer is down: losing the
        record would be worse than lacking the receipt. The receipt is issued
        afterwards with `POST /payments/{id}/receipt`.

        Allocations are refused if they exceed the payment, exceed what a
        document still owes, name a document belonging to another customer, or
        name a document that is not issued.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PaymentInput' }
      responses:
        '201':
          description: |
            Recorded. `Idempotent-Replay: true` means this was a replay of an
            earlier request under the same key.
          headers:
            Idempotent-Replay:
              schema: { type: boolean }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Payment' }
        '403': { $ref: '#/components/responses/Error' }
        '409':
          description: The idempotency key was used with a different request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '422': { $ref: '#/components/responses/Error' }
        '500':
          description: |
            The PDF renderer is unavailable, so no receipt could be issued and
            nothing was recorded. Retry with `issue_receipt: false`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  /payments/by-method:
    get:
      tags: [payments]
      summary: Payments summarized by method (plan.md 15)
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Totals
          content:
            application/json:
              schema:
                type: object
                properties:
                  totals:
                    type: array
                    items: { $ref: '#/components/schemas/MethodTotal' }
  /payments/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      tags: [payments]
      summary: One payment, with its allocations
      responses:
        '200':
          description: Payment
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Payment' }
        '404': { $ref: '#/components/responses/Error' }
  /payments/{id}/receipt:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    post:
      tags: [payments]
      summary: Issue the receipt for a payment that has none
      description: |
        For a payment recorded with `issue_receipt: false`. A payment has at most
        one receipt; issuing a second is refused. Once the receipt exists the
        payment row is frozen — a database trigger refuses any further change,
        because the receipt states its amount, method and date.
      responses:
        '200':
          description: The payment, now with its receipt attached
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Payment' }
        '409':
          description: The payment already has a receipt
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  /reports/revenue:
    get:
      tags: [reports]
      summary: Revenue, broken down as asked
      parameters:
        - name: by
          in: query
          schema:
            type: string
            enum: [period, customer, service, activity]
            default: period
        - name: granularity
          in: query
          description: Only for `by=period`
          schema:
            type: string
            enum: [day, month, year]
            default: month
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Rows
          content:
            application/json:
              schema:
                type: object
                properties:
                  rows:
                    type: array
                    items: { $ref: '#/components/schemas/RevenueRow' }
                  range:
                    type: object
                    properties:
                      from: { type: string }
                      to: { type: string }
        '422': { $ref: '#/components/responses/Error' }
...
  /reports/turnover:
    get:
      tags: [reports]
      summary: Progress against the exempt-dealer annual ceiling
      description: |
        A year with no configured ceiling reports `threshold_known: false`
        rather than a guessed number (plan.md 22.3).
      parameters:
        - name: year
          in: query
          schema: { type: integer, minimum: 2000, maximum: 2200 }
      responses:
        '200':
          description: Turnover
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Turnover' }
        '422': { $ref: '#/components/responses/Error' }
...
  /payments/preview:
    post:
      tags: [payments]
      summary: See the receipt a payment would produce, before recording it
      description: |
        A receipt is never a draft — it commits together with its payment — so
        this is the only chance to look at one before it becomes immutable.

        Takes the payment exactly as the operator has filled it in, validates it
        the same way recording would, and renders the receipt that would be
        issued. **Records nothing.** The preview and the real issuance build the
        receipt through the same code, so what is approved is what gets issued.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PaymentInput' }
      responses:
        '200':
          description: The preview, inline
          content:
            application/pdf:
              schema: { type: string, format: binary }
        '422': { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: |
        Set by `POST /auth/login`. Unsafe methods must also send
        `X-CSRF-Token`, echoing the `csrf_token` cookie.
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, maxLength: 200 }
      description: |
        Makes a retry safe. A repeat with the same key returns the original
        response; the same key with a different request is refused with 409.
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable English machine identifier
              examples: [UNAUTHORIZED, VALIDATION_FAILED, DOCUMENT_IMMUTABLE]
            message:
              type: string
              description: Hebrew text, shown to the user as-is
            details:
              type: object
              additionalProperties: true
              description: Field-level messages, keyed by field name
            request_id:
              type: string
...
    DocumentType:
      type: string
      enum: [TRANSACTION_INVOICE, PAYMENT_REQUEST, RECEIPT]
      description: |
        חשבונית עסקה, דרישת תשלום, קבלה. `TAX_INVOICE` (חשבונית מס) is absent by
        design: an exempt dealer may not issue one.
...
    PaymentMethod:
      type: string
      enum: [CASH, BANK_TRANSFER, CREDIT_CARD, BIT, PAYBOX, CHECK, OTHER]
    PaymentAllocationInput:
      type: object
      required: [document_id, amount_agorot]
      properties:
        document_id: { type: string, format: uuid }
        amount_agorot: { type: integer, format: int64, minimum: 1 }
    PaymentAllocation:
      allOf:
        - $ref: '#/components/schemas/PaymentAllocationInput'
        - type: object
          properties:
            full_number: { type: string, readOnly: true }
            document_type:
              allOf: [{ $ref: '#/components/schemas/DocumentType' }]
              readOnly: true
    PaymentInput:
      type: object
      required: [customer_id, amount_agorot, received_at, method]
      properties:
        customer_id: { type: string, format: uuid }
        amount_agorot:
          type: integer
          format: int64
          minimum: 1
          description: |
            Always positive. A refund is a different act with different paperwork
            and is not modelled as a negative payment.
        received_at:
          type: string
          format: date
          description: Business date, already interpreted in Asia/Jerusalem
        method: { $ref: '#/components/schemas/PaymentMethod' }
        reference:
          type: string
          description: Cheque number, transfer reference, last digits of a card
        notes: { type: string }
        activity_id: { type: [string, 'null'], format: uuid }
        allocations:
          type: array
          description: |
            What this payment settles. Omit it entirely for money on account.
            The sum may not exceed the payment, and no single allocation may
            exceed what its document still owes.
          items: { $ref: '#/components/schemas/PaymentAllocationInput' }
        issue_receipt:
          type: boolean
          default: true
          description: |
            Issue the receipt with the payment. Set false only when the renderer
            is unavailable; issue the receipt afterwards.
    Payment:
      type: object
      properties:
        id: { type: string, format: uuid }
        customer_id: { type: string, format: uuid }
        customer_name: { type: string }
        amount_agorot: { type: integer, format: int64 }
        received_at: { type: string, format: date }
        method: { $ref: '#/components/schemas/PaymentMethod' }
        method_hebrew: { type: string }
        reference: { type: string }
        notes: { type: string }
        activity_id: { type: [string, 'null'], format: uuid }
        activity_name: { type: [string, 'null'] }
        receipt_document_id:
          type: [string, 'null']
          format: uuid
          description: Null means no receipt has been issued yet
        receipt_number: { type: [string, 'null'] }
        allocated_agorot:
          type: integer
          format: int64
          description: How much of this payment settles specific documents
        on_account_agorot:
          type: integer
          format: int64
          description: The remainder, which settles nothing in particular
        allocations:
          type: array
          items: { $ref: '#/components/schemas/PaymentAllocation' }
        created_at: { type: string, format: date-time }
    PaymentPage:
      type: object
      properties:
        payments:
          type: array
          items: { $ref: '#/components/schemas/Payment' }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        total_amount_agorot:
          type: integer
          format: int64
          description: Sum of every payment matching the filter, not only this page
    MethodTotal:
      type: object
      properties:
        method: { $ref: '#/components/schemas/PaymentMethod' }
        method_hebrew: { type: string }
        count: { type: integer }
        total_agorot: { type: integer, format: int64 }
...
    RevenueRow:
      type: object
      properties:
        key: { type: string }
        label: { type: string }
        count: { type: integer }
        amount_agorot: { type: integer, format: int64 }
...
    Turnover:
      type: object
      properties:
        year: { type: integer }
        revenue_agorot: { type: integer, format: int64 }
        threshold_known:
          type: boolean
          description: False when no ceiling is configured for that year
        threshold_agorot: { type: integer, format: int64 }
        remaining_agorot: { type: integer, format: int64 }
        percent_used: { type: number }
...